Add expect_no_error and improve timeout (#1)
* Add `expect_no_error(message, function, exception=BaseException)`

Tests if the function, when run, does not raise any exception that can be caught by `exception`. Generates a failed assertion if not met.

* Improve `@timeout(sec)`

The wrapped function is again wrapped inside `expect_no_error`, so that any exception raised inside the child process can be caught right there. If this happens, generates a failed assertion.

* Fix formatting

Fixes Codewars/codewars-runner-cli#698